fix all this stuff
